Refactoring; UglifyJS plugin added
