+++ config/webpack.prod.js
@@ -1,3 +1,5 @@
+const UglifyJSPlugin = require('uglifyjs-webpack-plugin');
+
 module.exports = {
     entry: "./src/main.ts",
     output: {
@@ -17,4 +19,7 @@ module.exports = {
             { enforce: "pre", test: /\.js$/, loader: "source-map-loader" }
         ]
     },
+    plugins: [
+        new UglifyJSPlugin()
+    ]
 };
