+++ package.json
@@ -16,6 +16,7 @@
     "karma-typescript": "^3.0.5",
     "source-map-loader": "^0.2.1",
     "typescript": "^2.4.2",
+    "uglifyjs-webpack-plugin": "^1.1.2",
     "webpack": "^3.4.1"
   },
   "scripts": {
